+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.65.0] - 2020-12-02
+
 ### Added
 - `actionOnProductClick` prop to `ProductSummaryList`.